+++ jest.setup.js
@@ -26,10 +26,6 @@ jest.mock("@bugsnag/js", () => ({
   },
 }));
 
-jest.mock("./src/node-lib/curriculum-api", () =>
-  jest.requireActual("./src/node-lib/curriculum-api/__mocks__"),
-);
-
 jest.mock("./src/node-lib/curriculum-api-2023", () =>
   jest.requireActual("./src/node-lib/curriculum-api-2023/__mocks__"),
 );