+++ src/components/PupilComponents/QuizMCQMultiAnswerFeedback/QuizMCQMultiAnswerFeedback.test.tsx
@@ -8,7 +8,7 @@ import { createQuizEngineContext } from "../pupilTestHelpers/createQuizEngineCon
 
 import { QuizMCQMultiAnswerFeedback } from "./QuizMCQMultiAnswerFeedback";
 
-import { quizQuestions } from "@/node-lib/curriculum-api-2023/fixtures/quizElements.fixture";
+import { quizQuestions } from "@/node-lib/curriculum-api-2023/fixtures/quizElements.new.fixture";
 import renderWithTheme from "@/__tests__/__helpers__/renderWithTheme";
 
 const questionsArrayFixture = quizQuestions || [];