+++ src/node-lib/curriculum-api-2023/__mocks__/index.ts
@@ -1,18 +1,20 @@
-import { specialistSubjectListingFixture2023 } from "../fixtures/specialistSubjectListing.fixture";
+import unitListingFixture from "../fixtures/unitListing.fixture";
 
+import { specialistSubjectListingFixture2023 } from "@/node-lib/curriculum-api-2023/fixtures/specialistSubjectListing.fixture";
 import programmeListingFixture from "@/node-lib/curriculum-api-2023/fixtures/programmeListing.fixture";
 import lessonListingFixture from "@/node-lib/curriculum-api-2023/fixtures/lessonListing.fixture";
 import lessonOverviewFixture from "@/node-lib/curriculum-api-2023/fixtures/lessonOverview.fixture";
 import { CurriculumApi } from "@/node-lib/curriculum-api-2023";
 import subjectPhaseOptionsFixture from "@/node-lib/curriculum-api-2023/fixtures/subjectPhaseOptions.fixture";
 import { curriculumOverviewMVFixture } from "@/node-lib/curriculum-api-2023/fixtures/curriculumOverview.fixture";
-import { subjectListingFixture2023 } from "@/node-lib/curriculum-api/fixtures/subjectListing.fixture";
-import teachersHomePageFixture from "@/node-lib/curriculum-api/fixtures/teachersHomePage.fixture";
-import lessonDownloadsFixtures from "@/node-lib/curriculum-api/fixtures/lessonDownloads.fixture";
-import pupilLessonOverviewFixture from "@/node-lib/curriculum-api/fixtures/pupilLessonOverview.fixture";
+import teachersHomePageFixture from "@/node-lib/curriculum-api-2023/fixtures/teachersHomePage.fixture";
 import specialistUnitListingFixture from "@/components/TeacherViews/SpecialistUnitListing/SpecialistUnitListing.fixture";
 import specialistLessonOverviewFixture from "@/node-lib/curriculum-api-2023/fixtures/specialistLessonOverview.fixture";
+import { lessonContentFixture } from "@/node-lib/curriculum-api-2023/fixtures/lessonContent.fixture";
+import { lessonBrowseDataFixture } from "@/node-lib/curriculum-api-2023/fixtures/lessonBrowseData.fixture";
 import lessonShareFixtures from "@/node-lib/curriculum-api-2023/fixtures/lessonShare.fixture";
+import lessonDownloadsFixtures from "@/node-lib/curriculum-api-2023/fixtures/lessonDownloads.fixture";
+import { subjectListingFixture2023 } from "@/node-lib/curriculum-api-2023/fixtures/subjectListing.fixture";
 
 const curriculumApi: Pick<
   CurriculumApi,
@@ -24,14 +26,14 @@ const curriculumApi: Pick<
   | "teachersHomePage"
   | "lessonOverviewCanonical"
   | "lessonDownloadsCanonical"
-  | "pupilLessonOverview"
-  | "pupilLessonOverviewCanonical"
+  | "pupilLessonQuery"
   | "specialistSubjectListing"
   | "specialistUnitListing"
   | "specialistLessonOverviewCanonical"
   | "lessonShare"
   | "lessonOverview"
   | "lessonDownloads"
+  | "unitListing"
 > = {
   subjectPhaseOptions: jest.fn(async () => {
     return subjectPhaseOptionsFixture();
@@ -51,14 +53,14 @@ const curriculumApi: Pick<
   teachersHomePage: jest.fn(async () => {
     return teachersHomePageFixture();
   }),
-  pupilLessonOverview: jest.fn(async () => {
-    return pupilLessonOverviewFixture();
+  pupilLessonQuery: jest.fn(async () => {
+    return {
+      browseData: lessonBrowseDataFixture({}),
+      content: lessonContentFixture({}),
+    };
   }),
   lessonShare: jest.fn(async () => {
     return lessonShareFixtures();
-  }),
-  pupilLessonOverviewCanonical: jest.fn(async () => {
-    return pupilLessonOverviewFixture();
   }),
   lessonOverview: jest.fn(async () => {
     return lessonOverviewFixture();
@@ -77,6 +79,9 @@ const curriculumApi: Pick<
   }),
   lessonDownloads: jest.fn(async () => {
     return lessonDownloadsFixtures();
+  }),
+  unitListing: jest.fn(async () => {
+    return unitListingFixture();
   }),
   specialistSubjectListing: jest.fn(async () => {
     return {